requirements.txt
Pillow==10.1.0
pycodestyle==2.11.1
Flask==2.3.2